refactor bucket api
