+++ src/http_frontend/bucket_api.rs
@@ -3,7 +3,11 @@
 //    License, v. 2.0. If a copy of the MPL was not distributed with this
 //    file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+mod create;
 mod get;
+mod head;
+mod remove;
+mod update;
 
 use std::sync::{Arc, RwLock};
 
@@ -12,19 +16,23 @@ use axum::extract::{FromRequest, Path, State};
 use axum::headers::HeaderMapExt;
 use axum::http::{Request, StatusCode};
 use axum::response::{IntoResponse, Response};
+use axum::routing::{delete, get, head, post, put};
 use axum::{async_trait, headers};
 use bytes::Bytes;
 use hyper::HeaderMap;
 use serde_json::{json, Value};
 
 use crate::core::status::HttpError;
+use crate::http_frontend::bucket_api::create::create_bucket;
+use crate::http_frontend::bucket_api::get::get_bucket;
+use crate::http_frontend::bucket_api::head::head_bucket;
+use crate::http_frontend::bucket_api::remove::remove_bucket;
+use crate::http_frontend::bucket_api::update::update_bucket;
 use crate::http_frontend::middleware::check_permissions;
 use crate::http_frontend::HttpServerState;
 use crate::storage::proto::bucket_settings::QuotaType;
 use crate::storage::proto::BucketSettings;
 use crate::storage::proto::FullBucketInfo;
-
-pub struct BucketApi {}
 
 impl IntoResponse for FullBucketInfo {
     fn into_response(self) -> Response {
@@ -105,72 +113,13 @@ where
     }
 }
 
-impl BucketApi {
-    // GET /b/:bucket_name
-    pub async fn get_bucket(
-        State(components): State<Arc<RwLock<HttpServerState>>>,
-        Path(bucket_name): Path<String>,
-        headers: HeaderMap,
-    ) -> Result<FullBucketInfo, HttpError> {
-        check_permissions(Arc::clone(&components), headers, AuthenticatedPolicy {})?;
-        let mut components = components.write().unwrap();
-        components.storage.get_bucket(&bucket_name)?.info()
-    }
-
-    // HEAD /b/:bucket_name
-    pub async fn head_bucket(
-        State(components): State<Arc<RwLock<HttpServerState>>>,
-        Path(bucket_name): Path<String>,
-        headers: HeaderMap,
-    ) -> Result<(), HttpError> {
-        check_permissions(Arc::clone(&components), headers, AuthenticatedPolicy {})?;
-        let mut components = components.write().unwrap();
-        components.storage.get_bucket(&bucket_name)?;
-        Ok(())
-    }
-
-    // POST /b/:bucket_name
-    pub async fn create_bucket(
-        State(components): State<Arc<RwLock<HttpServerState>>>,
-        Path(bucket_name): Path<String>,
-        headers: HeaderMap,
-        settings: BucketSettings,
-    ) -> Result<(), HttpError> {
-        check_permissions(Arc::clone(&components), headers, FullAccessPolicy {})?;
-        let mut components = components.write().unwrap();
-        components
-            .storage
-            .create_bucket(&bucket_name, settings.into())?;
-        Ok(())
-    }
-
-    // PUT /b/:bucket_name
-    pub async fn update_bucket(
-        State(components): State<Arc<RwLock<HttpServerState>>>,
-        Path(bucket_name): Path<String>,
-        headers: HeaderMap,
-        settings: BucketSettings,
-    ) -> Result<(), HttpError> {
-        check_permissions(Arc::clone(&components), headers, FullAccessPolicy {})?;
-        let mut components = components.write().unwrap();
-        let bucket = components.storage.get_bucket(&bucket_name)?;
-        bucket.set_settings(settings.into())
-    }
-
-    // DELETE /b/:bucket_name
-    pub async fn remove_bucket(
-        State(components): State<Arc<RwLock<HttpServerState>>>,
-        Path(bucket_name): Path<String>,
-        headers: HeaderMap,
-    ) -> Result<(), HttpError> {
-        check_permissions(Arc::clone(&components), headers, FullAccessPolicy {})?;
-        let mut components = components.write().unwrap();
-        components.storage.remove_bucket(&bucket_name)?;
-        components
-            .token_repo
-            .remove_bucket_from_tokens(&bucket_name)?;
-        Ok(())
-    }
+pub fn create_bucket_api_routes() -> axum::Router<Arc<RwLock<HttpServerState>>> {
+    axum::Router::new()
+        .route("/:bucket_name", get(get_bucket))
+        .route("/:bucket_name", head(head_bucket))
+        .route("/:bucket_name", post(create_bucket))
+        .route("/:bucket_name", put(update_bucket))
+        .route("/:bucket_name", delete(remove_bucket))
 }
 
 #[cfg(test)]
@@ -185,73 +134,9 @@ mod tests {
 
     use axum::http::Method;
     use hyper::Body;
+    use rstest::fixture;
     use std::path::PathBuf;
     use std::sync::{Arc, RwLock};
-
-    #[tokio::test]
-    async fn test_get_bucket() {
-        let components = setup();
-        let info = BucketApi::get_bucket(
-            State(components),
-            Path("bucket-1".to_string()),
-            HeaderMap::new(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(info.info.unwrap().name, "bucket-1");
-    }
-
-    #[tokio::test]
-    async fn test_head_bucket() {
-        let components = setup();
-        BucketApi::head_bucket(
-            State(components),
-            Path("bucket-1".to_string()),
-            HeaderMap::new(),
-        )
-        .await
-        .unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_create_bucket() {
-        let components = setup();
-        BucketApi::create_bucket(
-            State(components),
-            Path("bucket-2".to_string()),
-            HeaderMap::new(),
-            BucketSettings::default(),
-        )
-        .await
-        .unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_update_bucket() {
-        let components = setup();
-        BucketApi::update_bucket(
-            State(components),
-            Path("bucket-1".to_string()),
-            HeaderMap::new(),
-            BucketSettings::default(),
-        )
-        .await
-        .unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_remove_bucket() {
-        {
-            let components = setup();
-            BucketApi::remove_bucket(
-                State(components),
-                Path("bucket-1".to_string()),
-                HeaderMap::new(),
-            )
-            .await
-            .unwrap();
-        }
-    }
 
     #[tokio::test]
     async fn test_bucket_settings_quota_parsing() {
@@ -271,24 +156,5 @@ mod tests {
                 .unwrap(),
             "Failed to parse quota type: 1"
         );
-    }
-
-    fn setup() -> Arc<RwLock<HttpServerState>> {
-        let data_path = tempfile::tempdir().unwrap().into_path();
-
-        let mut components = HttpServerState {
-            storage: Storage::new(PathBuf::from(data_path.clone())),
-            auth: TokenAuthorization::new(""),
-            token_repo: create_token_repository(data_path.clone(), ""),
-            console: ZipAssetManager::new(&[]),
-            base_path: "/".to_string(),
-        };
-
-        components
-            .storage
-            .create_bucket("bucket-1", BucketSettings::default())
-            .unwrap();
-
-        Arc::new(RwLock::new(components))
     }
 }
