+++ src/http_frontend/server_api.rs
@@ -61,44 +61,3 @@ pub fn create_server_api_routes() -> axum::Router<Arc<RwLock<HttpServerState>>> 
         .route("/info", get(info::info))
         .route("/alive", head(|| async { StatusCode::OK }))
 }
-
-#[cfg(test)]
-mod tests {
-    use std::path::PathBuf;
-
-    use hyper::http::request::Builder;
-    use rstest::fixture;
-
-    use crate::asset::asset_manager::ZipAssetManager;
-    use crate::auth::token_auth::TokenAuthorization;
-    use crate::auth::token_repository::create_token_repository;
-    use crate::storage::proto::BucketSettings;
-    use crate::storage::storage::Storage;
-
-    use super::*;
-
-    #[fixture]
-    pub(crate) fn components() -> Arc<RwLock<HttpServerState>> {
-        let data_path = tempfile::tempdir().unwrap().into_path();
-
-        let mut components = HttpServerState {
-            storage: Storage::new(PathBuf::from(data_path.clone())),
-            auth: TokenAuthorization::new(""),
-            token_repo: create_token_repository(data_path.clone(), ""),
-            console: ZipAssetManager::new(&[]),
-            base_path: "/".to_string(),
-        };
-
-        components
-            .storage
-            .create_bucket("bucket-1", BucketSettings::default())
-            .unwrap();
-        components
-            .storage
-            .create_bucket("bucket-2", BucketSettings::default())
-            .unwrap();
-
-        let _req = Builder::new().body(()).unwrap();
-        Arc::new(RwLock::new(components))
-    }
-}
