+++ src/http_frontend/token_api.rs
@@ -96,58 +96,3 @@ pub fn create_token_api_routes() -> axum::Router<Arc<RwLock<HttpServerState>>> {
         .route("/:token_name", get(get_token))
         .route("/:token_name", delete(remove_token))
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::asset::asset_manager::ZipAssetManager;
-    use crate::auth::token_auth::TokenAuthorization;
-    use crate::auth::token_repository::create_token_repository;
-    use crate::storage::storage::Storage;
-
-    use crate::storage::proto::BucketSettings;
-
-    use axum::headers::Authorization;
-    use rstest::fixture;
-    use std::path::PathBuf;
-
-    #[fixture]
-    pub(crate) fn components() -> Arc<RwLock<HttpServerState>> {
-        let data_path = tempfile::tempdir().unwrap().into_path();
-
-        let mut components = HttpServerState {
-            storage: Storage::new(PathBuf::from(data_path.clone())),
-            auth: TokenAuthorization::new("inti-token"),
-            token_repo: create_token_repository(data_path.clone(), "init-token"),
-            console: ZipAssetManager::new(&[]),
-            base_path: "/".to_string(),
-        };
-
-        components
-            .storage
-            .create_bucket("bucket-1", BucketSettings::default())
-            .unwrap();
-        components
-            .storage
-            .create_bucket("bucket-2", BucketSettings::default())
-            .unwrap();
-
-        let permissions = Permissions {
-            read: vec!["bucket-1".to_string(), "bucket-2".to_string()],
-            ..Default::default()
-        };
-        components
-            .token_repo
-            .create_token("test", permissions)
-            .unwrap();
-
-        Arc::new(RwLock::new(components))
-    }
-
-    #[fixture]
-    pub(crate) fn headers() -> HeaderMap {
-        let mut headers = HeaderMap::new();
-        headers.typed_insert(Authorization::bearer("init-token").unwrap());
-        headers
-    }
-}
