+++ src/http_frontend/token_api/create.rs
@@ -30,7 +30,7 @@ pub async fn create_token(
 mod tests {
     use super::*;
 
-    use crate::http_frontend::token_api::tests::{components, headers};
+    use crate::http_frontend::tests::{components, headers};
 
     use rstest::rstest;
