+++ src/http_frontend/token_api/get.rs
@@ -28,7 +28,7 @@ pub async fn get_token(
 mod tests {
     use super::*;
 
-    use crate::http_frontend::token_api::tests::{components, headers};
+    use crate::http_frontend::tests::{components, headers};
 
     use rstest::rstest;
