+++ src/http_frontend/token_api/remove.rs
@@ -26,10 +26,8 @@ pub async fn remove_token(
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    use crate::http_frontend::bucket_api::BucketApi;
+    use crate::http_frontend::tests::{components, headers};
     use crate::http_frontend::token_api::get::get_token;
-    use crate::http_frontend::token_api::tests::{components, headers};
 
     use rstest::rstest;
 
@@ -38,40 +36,5 @@ mod tests {
     async fn test_remove_token(components: Arc<RwLock<HttpServerState>>, headers: HeaderMap) {
         let token = remove_token(State(components), Path("test".to_string()), headers).await;
         assert!(token.is_ok());
-    }
-
-    #[rstest]
-    #[tokio::test]
-    async fn test_remove_bucket_from_permission(
-        components: Arc<RwLock<HttpServerState>>,
-        headers: HeaderMap,
-    ) {
-        let token = get_token(
-            State(Arc::clone(&components)),
-            Path("test".to_string()),
-            headers.clone(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(
-            token.permissions.unwrap().read,
-            vec!["bucket-1".to_string(), "bucket-2".to_string()]
-        );
-
-        BucketApi::remove_bucket(
-            State(Arc::clone(&components)),
-            Path("bucket-1".to_string()),
-            headers.clone(),
-        )
-        .await
-        .unwrap();
-
-        let token = get_token(State(components), Path("test".to_string()), headers)
-            .await
-            .unwrap();
-        assert_eq!(
-            token.permissions.unwrap().read,
-            vec!["bucket-2".to_string()]
-        );
     }
 }
